Update: Move command classes to seperate file and into Command Namespace
