+++ Makefile
@@ -3,10 +3,10 @@
 #                                                         :::      ::::::::    #
 #    Makefile                                           :+:      :+:    :+:    #
 #                                                     +:+ +:+         +:+      #
-#    By: orezek <orezek@student.42.fr>              +#+  +:+       +#+         #
+#    By: mbartos <mbartos@student.42prague.com>     +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/11/21 10:59:06 by mbartos           #+#    #+#              #
-#    Updated: 2024/09/30 21:07:41 by orezek           ###   ########.fr        #
+#    Updated: 2024/10/18 11:49:43 by mbartos          ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -16,44 +16,48 @@ RED =	\033[31m
 BCYAN =	\033[96m
 NC =	\033[0m
 
-# Name of the executable
 NAME =		ircserv
 
-# Compiler and flags
 CCP =		c++
 CFLAGS =	-g -std=c++98 \
-			-I./include/
+			-I./include/ -I./include/commands
 
-# Source and object paths
 SRC_PATH = ./src/
-SRC =		$(wildcard $(SRC_PATH)*.cpp)
+COMMANDS_PATH = ./src/commands/
+
+SRC =		$(wildcard $(SRC_PATH)*.cpp) $(wildcard $(COMMANDS_PATH)*.cpp)
 
 # Objects and dependencies
 OBJ_PATH =	obj/
-OBJ =		$(SRC:$(SRC_PATH)%.cpp=$(OBJ_PATH)%.o)  # Ensures correct object file paths
+OBJ_COMMANDS_PATH = $(OBJ_PATH)commands/
+OBJ =		$(patsubst $(SRC_PATH)%.cpp,$(OBJ_PATH)%.o,$(wildcard $(SRC_PATH)*.cpp)) \
+			$(patsubst $(COMMANDS_PATH)%.cpp,$(OBJ_COMMANDS_PATH)%.o,$(wildcard $(COMMANDS_PATH)*.cpp))
+
 DEPS =		$(OBJ:.o=.d)
 
-# Binary directory
 BIN_PATH = bin/
 
 all: $(BIN_PATH)$(NAME)
 
-# Rule to compile .cpp files to .o files
+# Rule to compile general .cpp files into obj/
 $(OBJ_PATH)%.o: $(SRC_PATH)%.cpp | $(OBJ_PATH)
 	$(CCP) $(CFLAGS) -MMD -MP -c $< -o $@
 
-# Include dependency files
+# Rule to compile command .cpp files into obj/commands/
+$(OBJ_COMMANDS_PATH)%.o: $(COMMANDS_PATH)%.cpp | $(OBJ_COMMANDS_PATH)
+	$(CCP) $(CFLAGS) -MMD -MP -c $< -o $@
+
 -include $(DEPS)
 
-# Create object directory if it doesn't exist
 $(OBJ_PATH):
 	@mkdir -p $(OBJ_PATH)
 
-# Create binary directory if it doesn't exist
+$(OBJ_COMMANDS_PATH):
+	@mkdir -p $(OBJ_COMMANDS_PATH)
+
 $(BIN_PATH):
 	@mkdir -p $(BIN_PATH)
 
-# Link object files to create the final executable
 $(BIN_PATH)$(NAME): $(OBJ) | $(BIN_PATH)
 	@echo "$(BOLD)$(BCYAN)[ Compiling $(NAME)... ]$(NC)"
 	$(CCP) $(CFLAGS) $(OBJ) -o $@
