+++ include/commands/NickCommand.hpp
@@ -6,7 +6,7 @@
 /*   By: mbartos <mbartos@student.42prague.com>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/25 16:20:19 by mbartos           #+#    #+#             */
-/*   Updated: 2024/10/16 12:40:51 by mbartos          ###   ########.fr       */
+/*   Updated: 2024/10/18 11:12:11 by mbartos          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,6 +16,9 @@
 #include "ClientMessage.hpp"
 #include "ServerDataManager.hpp"
 #include "ServerResponse.hpp"
+
+namespace Commands
+{
 
 class NickCommand
 {
@@ -46,3 +49,5 @@ class NickCommand
 		void setServerResponseValid(User* user);
 		void addServerResponseToClient();
 };
+
+}  // namespace Commands
