+++ src/commands/PassCommand.cpp
@@ -1,94 +1,96 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   UserCommand.cpp                                    :+:      :+:    :+:   */
+/*   PassCommand.cpp                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: mbartos <mbartos@student.42prague.com>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/09/30 14:13:45 by mbartos           #+#    #+#             */
-/*   Updated: 2024/10/16 12:56:18 by mbartos          ###   ########.fr       */
+/*   Created: 2024/09/30 13:05:16 by mbartos           #+#    #+#             */
+/*   Updated: 2024/10/18 11:19:37 by mbartos          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "UserCommand.hpp"
+#include "PassCommand.hpp"
 
-UserCommand::UserCommand(Client* client, ClientMessage& clientMessage) : client(client), serverData(ServerDataManager::getInstance()), clientMessage(clientMessage) /*, user(NULL)*/
+namespace Commands
+{
+
+PassCommand::PassCommand(Client* client, ClientMessage& clientMessage) : client(client), serverData(ServerDataManager::getInstance()), clientMessage(clientMessage)
 {
 	if (client->user.isValidServerUser() == true)
 	{
 		this->setServerResponse462();  // user already validated
-		this->addServerResponseToClient();
 		return;
 	}
 
-	std::string username = clientMessage.getParameterAtPosition(0);
-	std::string hostname = clientMessage.getParameterAtPosition(1);
-	std::string servername = clientMessage.getParameterAtPosition(2);
-	std::string realname = clientMessage.getParameterAtPosition(3);
+	client->user.setPassSent(true);
 
-	if (realname.empty() || servername.empty() || hostname.empty() || username.empty())
+	std::string passedPassword = this->clientMessage.getFirstParameter();
+	std::string serverPassword = this->serverData.getServerPassword();
+	if (passedPassword.empty())
 	{
 		this->setServerResponse461();
-		this->addServerResponseToClient();
 		return;
 	}
-
-	// check parameters, if they are valid
-
-	client->user.setUsername(username);
-	client->user.setHostname(hostname);
-	client->user.setServername(servername);
-	client->user.setRealname(realname);
-	client->user.setUserValid(true);
+	if (passedPassword == serverPassword)
+	{
+		client->user.setPassValid(true);
+		this->setServerResponseValid();
+	}
+	// else if
+	// {
+	// 	/* Pass not valid */
+	// }
 }
 
-UserCommand::UserCommand(UserCommand const& refObj) : client(refObj.client), serverData(refObj.serverData), clientMessage(refObj.clientMessage), serverResponse(refObj.serverResponse) /*, user(refObj.user)*/ {}
+PassCommand::~PassCommand() {}
 
-UserCommand& UserCommand::operator=(UserCommand const& refObj)
+PassCommand::PassCommand(PassCommand const& refObj) : client(refObj.client), serverData(refObj.serverData), clientMessage(refObj.clientMessage), serverResponse(refObj.serverResponse) {}
+
+PassCommand& PassCommand::operator=(PassCommand const& refObj)
 {
 	if (this != &refObj)
 	{
 		this->client = refObj.client;
-		this->clientMessage = refObj.clientMessage;
 		// this->serverData = refObj.serverData;
+		this->clientMessage = refObj.clientMessage;
 		this->serverResponse = refObj.serverResponse;
-		// this->user = refObj.user;
 	}
 	return (*this);
 }
 
-UserCommand::~UserCommand() {}
-
-ServerResponse UserCommand::getServerResponse()
+ServerResponse PassCommand::getServerResponse()
 {
 	return (this->serverResponse);
 }
 
 // ---- PRIVATE ----
 
-void UserCommand::addServerResponseToClient()
+void PassCommand::addServerResponseToClient()
 {
 	client->serverResponses.push_back(serverResponse);
 }
 
-void UserCommand::setServerResponse461()
+void PassCommand::setServerResponse461()
 {
 	std::string nickname = client->user.getNickname();
 	if (nickname.empty())
 	{
 		nickname = "*";
 	}
+
 	std::string response = ":";
 	response.append(serverData.getServerName());
 	response.append(" 461 ");
 	response.append(nickname);
-	response.append(" NICK :Not enough parameters.\r\n");
+	response.append(" PASS :Not enough parameters.\r\n");
 	serverResponse.setAction(ServerResponse::SEND);
 	serverResponse.setResponse(response);
 	serverResponse.setClientsToSend(clientMessage.getFromUserFd());
+	this->addServerResponseToClient();
 }
 
-void UserCommand::setServerResponse462()
+void PassCommand::setServerResponse462()
 {
 	std::string nickname = client->user.getNickname();
 	if (nickname.empty())
@@ -104,4 +106,16 @@ void UserCommand::setServerResponse462()
 	serverResponse.setAction(ServerResponse::SEND);
 	serverResponse.setResponse(response);
 	serverResponse.setClientsToSend(clientMessage.getFromUserFd());
+	this->addServerResponseToClient();
 }
+
+void PassCommand::setServerResponseValid()
+{
+	// std::string response = "";
+
+	// serverResponse.setAction(ServerResponse::NOSEND);
+	// serverResponse.setResponse(response);
+	// serverResponse.setClientsToSend(clientMessage.getFromUserFd());
+}
+
+}  // namespace Commands
